Work around for pages not updating when refreshed.
This is a bug with parse_date_string in mongoose returning incorrect
timestamps. The easiest thing to do is port back differences in the new
mongoose server, but this doesn't fix parse_date_string only works
around it. I tried timegm but it is not available on windows, and
porting that uses gmtime which is only available in a thread-unsafe
form.